[feat/daengle-91] 댕글미터 도메인 설계

+++ daengle-domain/src/main/java/ddog/domain/groomer/GroomerDaengleMeter.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class GroomerDaengleMeter {
     private Long groomerDaengleMeterId;
     private Long groomerId;
-    private Long score;
+    private Integer score;
     private Long evaluationCount;
 }
